QOL stuff - remove platform requirements from `Package.swift` - fix incorrect use of inline array span when calling `wait` in `Epoll` - changed `HTTPResponseStatus.Code` from `Int` to `UInt16` - removed static computed property `count` from `InlineArrayProtocol` - `Router` updates - `DynamicRouteResponderProtocol` updates - added two compiled responder storages - moved some files around

+++ Sources/DestinyBlueprint/HTTPResponseStatus.swift
@@ -14,7 +14,7 @@ import SwiftSyntax
 /// - Standard Registry: https://www.iana.org/assignments/http-status-codes/http-status-codes.xhtml
 /// - Wikipedia: https://en.wikipedia.org/wiki/List_of_HTTP_status_codes
 public enum HTTPResponseStatus {
-    public typealias Code = Int
+    public typealias Code = UInt16
 }
 
 // MARK: Storage
